[zlib] osx fixes

+++ zlib/build.sh
@@ -2,5 +2,7 @@
 make -j9
 make install
 
-sed -i -r 's:\<(O[FN])\>:_Z_\1:g' "${PREFIX}"/include/*.h
-find "${PREFIX}" -name "libz.*a" -delete
+if [[ -z ${OSX_ARCH} ]] ; then
+   sed -i -r 's:\<(O[FN])\>:_Z_\1:g' "${PREFIX}"/include/*.h
+   find "${PREFIX}" -name "libz.*a" -delete
+fi
